fixed bug

+++ backend/src/main/java/ua/meetuply/backend/controller/BlogController.java
@@ -6,6 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import ua.meetuply.backend.model.BlogComment;
 import ua.meetuply.backend.model.BlogPost;
+import ua.meetuply.backend.service.AppUserService;
 import ua.meetuply.backend.service.BlogCommentService;
 import ua.meetuply.backend.service.BlogPostService;
 
@@ -20,6 +21,9 @@ public class BlogController {
 
     @Autowired
     BlogPostService blogPostService;
+
+    @Autowired
+    AppUserService appUserService;
 
     @GetMapping()
     public @ResponseBody Iterable<BlogPost> getAllBlogPosts() {
@@ -56,6 +60,8 @@ public class BlogController {
     @PutMapping("/{post-id}")
     public ResponseEntity updateBlogPost(@PathVariable("post-id") Integer blogPostId,
                                                    @RequestBody BlogPost blogPost) {
+        if (!(appUserService.getCurrentUser().getRole().getRoleName()=="admin"))
+            return ResponseEntity.badRequest().build();
         if (blogPostService.getBlogPostById(blogPostId) == null) {
             return ResponseEntity.badRequest().build();
         }
@@ -66,6 +72,8 @@ public class BlogController {
 
     @DeleteMapping("/{post-id}")
     public ResponseEntity deleteBlogPost(@PathVariable("post-id") Integer blogPostId){
+        if (!(appUserService.getCurrentUser().getRole().getRoleName()=="admin"))
+            return ResponseEntity.badRequest().build();
         if (blogPostService.getBlogPostById(blogPostId) == null) {
             return ResponseEntity.badRequest().build();
         }
@@ -106,6 +114,8 @@ public class BlogController {
     @PutMapping("/comments/{comment-id}")
     public ResponseEntity updateBlogComment(@PathVariable("comment-id") Integer blogCommentId,
                                                       @RequestBody BlogComment blogComment) {
+        if (!(appUserService.getCurrentUser().getRole().getRoleName()=="admin"))
+            return ResponseEntity.badRequest().build();
         if (blogCommentService.getBlogCommentById(blogCommentId) == null) {
             return ResponseEntity.badRequest().build();
         }
@@ -116,8 +126,10 @@ public class BlogController {
 
     @DeleteMapping("/comments/{comment-id}")
     public ResponseEntity deleteBlogComment(@PathVariable("comment-id") Integer blogCommentId){
+        if (!(appUserService.getCurrentUser().getRole().getRoleName()=="admin"))
+            return ResponseEntity.badRequest().build();
         if (blogCommentService.getBlogCommentById(blogCommentId) == null) {
-            ResponseEntity.badRequest().build();
+            return ResponseEntity.badRequest().build();
         }
         blogCommentService.deleteBlogComment(blogCommentId);
         return ResponseEntity.ok().build();
